requirements.txt
altair==4.1.0
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto3==1.14.9
botocore==1.17.9
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
docutils==0.15.2
en-core-web-md==2.3.0
entrypoints==0.3
enum-compat==0.0.3
heatmapz==0.0.4
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pipreqs==0.4.10
plac==1.1.3
plotly==4.8.1
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pydeck==0.4.0b2
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
scipy==1.5.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
spacy==2.3.0
srsly==1.0.2
streamlit==0.62.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
wasabi==0.7.0
watchdog==0.10.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarg==0.1.9
zipp==3.1.0
